+++ c/sort.h
@@ -55,6 +55,12 @@ lexicographic_less_3d(const double * restrict a, const double * restrict b)
 }
 
 static inline bool
+lexicographic_less_2d(const double * restrict a, const double * restrict b)
+{
+    return a[1] < b[1] || (a[1] == b[1] && a[0] <= b[0]);
+}
+
+static inline bool
 all_equal_double(const double * restrict a, const double * restrict b, dimension_t dim)
 {
     ASSUME(dim >= 2);
@@ -110,9 +116,8 @@ cmp_pdouble_asc_rev_2d(const void * restrict pa, const void * restrict pb)
 {
     const double * restrict a = (const double *) pa;
     const double * restrict b = (const double *) pb;
-    const double ax = a[0], ay = a[1], bx = b[0], by = b[1];
-    int cmpx = cmp_double_asc(ax, bx);
-    int cmpy = cmp_double_asc(ay, by);
+    int cmpx = cmp_double_asc(a[0], b[0]);
+    int cmpy = cmp_double_asc(a[1], b[1]);
     return cmpy ? cmpy : cmpx;
 }
 static inline int